// workers-backend/src/types.ts
export type ReposSchema = Repository[];
// ...
export type UserDetails = {
  name: string;
  html_url: string;
  avatar_url: string;
};

export type ProjectDetailsType = {
  name: string;
  description: string | null;
  fork: boolean;
  html_url: string;
  live_url: string | null;
  imagesKeysList : string[]
};